Support workspace trust and modification settings
VSCode v1.56+ changed how modifications made by the workspace are handled.  There is a setting for enabling modificaitons from the workspace, but this is planned to be deprecated in favor of the workspace trust system.  Update the readme with images and instructions for enabling both of these features.  When workspace trust takes over, the workspace modification setting can be removed and the setup should still work with minimal effort.

This should be a cleaner way to overwrite the Path variable for the integrated terminal without having to create a custom terminal profile.  A custom terminal profile is an option, but I still would need to enable workspace trust to set it as the default.  So the choice has been made to stick with the terminal.integrated.env setting for overwriting the path variable in case users want to create custom profiles, as the environment variables set here should persist across all profiles and terminal sessions.