Re-do launch configuration to use cpptools built-in remote config
There was a long-standing bug (https://github.com/Microsoft/vscode-cpptools/issues/833) in the debugger config that is resolved by this commit.  Pausing would not work with the previous configuration, which used setupCommands to script the GDB setup and used OpenOCD from within a pipe connection.  The cpptools extension did not know it was connected to a remote server, which caused pausing, setting breakpoints, and all sorts of other weird behavior to occur.

Launch.json no longer launches OpenOCD from a pipe connection.  Instead, it uses the new debugServerPath, debugServerArgs, miDebuggerServerAddress, and serverStarted config options.  This tells the cpptools extension how to launch OpenOCD as a server, which gets us the same nice all-in-one launching as the previous method.  It should be noted that filterStderr is necessary because OpenOCD outputs to stderr by default.  Now, the debugger works perfectly.

Flash was removed as a pre-launch task as well in favor of resetting the target micro on startup.  The user may not want to re-flash to program every time.