Manually set the PC after reset
Manually set the Program Counter to the Reset_Handler after reset.  This
enables smooth debugging on parts with a secure bootloader.  If a secure
part is reset, it enters into the secure bootloader verification in ROM,
which will prevent unsigned applications we've flashed from running
after reset.  However, the SWD port is always held open, so we can move
the program counter into the Reset_Handler manually to work around this.
This enables debugging the flashed application on secure parts for
development.

Additionally, this should be a transparent change for the non-secure
parts, which don't have any ROM.  A traditional reset will place the program
counter into the Reset_Handler anyways.